+++ gnn_model/pipeline.py
@@ -1,15 +1,20 @@
 import pandas as pd
+import torch
+from simulations.n_body_trajectory import n_body_simulation, generate_random_positions, generate_random_velocities, generate_unique_masses
+from gnn_model.node_data_list import node_data_list 
+from gnn_model.GNN_MLP import GNN_MLP
+from gnn_model.train_model import train_model
 
-def pipeline(train_iterations=100, test_iterations=40,
-                 N_train=2, N_test_list=[3, 4, 5, 6], T=500, dt=0.01, dim=2, hidden_channels=128,
-                 m_dim=2, out_channels=2, epochs=200, lr=0.001, save=False, model=None, training=True, testing=True):
+def pipeline(train_iterations=100, test_iterations=20,
+                 N_train=2, N_test_list=[2, 3, 4, 5, 6], T=500, dt=0.01, dim=2, hidden_channels=128,
+                 m_dim=2, out_channels=2, epochs=100, lr=0.001, save=False, model=None, training=True, testing=True, G=0.5, single_node=False):
     
     train_messages=None
     test_messages_all=None
     
     if training:
         # 1) Run training simulations with N_train
-        train_trajectories = [n_body_simulation(N=N_train, T=T, dt=dt, dim=dim, box_size=10, min_dist=2.5) for _ in range(train_iterations)]
+        train_trajectories = [n_body_simulation(N=N_train, T=T, dt=dt, dim=dim, box_size=10, min_dist=2.5, G=0.5) for _ in range(train_iterations)]
     
         train_graph_data = []
         for traj in train_trajectories:
@@ -52,6 +57,8 @@ def pipeline(train_iterations=100, test_iterations=40,
     # 7) Run and store test messages for each N in N_test_list
             test_messages_all = {}
             for N_test in N_test_list:
+                criterion = torch.nn.MSELoss()
+                total_loss = 0
                 test_trajectories = [n_body_simulation(N=N_test, T=T, dt=dt, dim=dim, box_size=30, min_dist=7) for _ in range(test_iterations)]
                 test_graph_data = []
                 for traj in test_trajectories:
@@ -60,7 +67,14 @@ def pipeline(train_iterations=100, test_iterations=40,
 
                 model.message_storage = []
                 for data in test_graph_data:
-                    _ = model(data.x, data.edge_index, save_messages=True)
+                    out = model(data.x, data.edge_index, save_messages=True)
+                    loss = criterion(out, data.y)
+                    total_loss +=loss.item()
+                print(len(test_graph_data))    
+
+                avg_loss = total_loss/len(test_graph_data)    
+
+                print(f"average loss per/over timestep N={N_test}:   {avg_loss}")
 
                 test_messages = pd.DataFrame(model.message_storage)
                 test_messages[['pos_i_x', 'pos_i_y']] = pd.DataFrame(test_messages['pos_i'].tolist())
